Update big-bird-non-sunset-rolls-mnk.json
added compass rose and moon & DC weapons, added/changed rolls for slugs, moon and DC weapons, Felwinter's Lie & Found Verdict, ARs
Prior to this I added quite a few changes at different times. So this one is for the new DC and Moon weapons but the other before this were other roll changes if they were not in the wishlist already.